Modified phone number types
Changed phone numbers from int to long as it was causing issues. Found out through testing in my python MVC that once you enter a phone number with a greater area code than 214, it would cause an error of using an invalid type, due to the size limitation of an int.

+++ ContosoPizza/Services/CustomerService.cs
@@ -33,7 +33,7 @@ public static class CustomerService {
     public static List<Customer> GetAllCustomers() => CustomerList;
     public static Customer? GetCustomer(string name) => CustomerList.FirstOrDefault(c => c.Name == name);
 
-    public static Customer? GetCustomer(int number) => CustomerList.FirstOrDefault(c => c.PhoneNumber == number);
+    public static Customer? GetCustomer(long number) => CustomerList.FirstOrDefault(c => c.PhoneNumber == number);
     //UPDATE
     public static void Update(Customer customer) {
         string token = GetToken(currentToken);
